SRSV2 block AQ (FR2 rev7-9), bench JSON layers, docs
- Add block_aq module (bounded qp_delta, variance-driven deltas, tests).
- Wire rev7 intra / rev8-9 P payloads, payload_kind, adaptive_quant block_wire stats.
- P decode: treat rev8 like rev4 for adaptive residual chunks (entropy_chunks).
- bench_srsv2: nest frame_aq (MB activity) vs block_aq_wire (on-wire 8x8); Markdown sections.
- Docs: frame vs block AQ semantics; intra U/V deltas vs P luma-only; codec overview.

+++ crates/libsrs_video/src/srsv2/subpel.rs
@@ -219,7 +219,7 @@ mod tests {
         let mut cur = VideoPlane::<u8>::try_new(w, w, w as usize).unwrap();
         for y in 0..16 {
             for x in 0..16 {
-                let v = sample_luma_bilinear_qpel(&refp, x as i32, y as i32, -2, 0);
+                let v = sample_luma_bilinear_qpel(&refp, x, y, -2, 0);
                 cur.samples[y as usize * cur.stride + x as usize] = v;
             }
         }
